requirements.txt
torch>=2.0.0
torchvision>=0.15.0
# PyTorch wheels are often built against NumPy 1.x; NumPy 2.x triggers ABI warnings/errors.
numpy>=1.24.0,<2
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tifffile>=2023.7.0
tqdm>=4.65.0
PyYAML>=6.0
kagglehub>=0.3.0
